+++ workers/unity/Packages/io.improbable.gdk.playerlifecycle/PlayerLifecycleHelper.cs
@@ -33,17 +33,15 @@ namespace Improbable.Gdk.PlayerLifecycle
                 throw new InvalidOperationException("Provided World does not have an associated worker");
             }
 
-            if (entitySystem.GetEntitiesInView().Contains(entityId))
+            if (!entitySystem.GetEntitiesInView().Contains(entityId))
             {
-                throw new InvalidOperationException(
-                    $"Entity with SpatialOS Entity ID {entityId.Id} is not in this worker's view");
+                return false;
             }
 
             if (!updateSystem.HasComponent(OwningWorker.ComponentId, entityId))
             {
                 return false;
             }
-
 
             var ownerId = updateSystem.GetComponent<OwningWorker.Snapshot>(entityId).WorkerId;
             return worker.WorkerId == ownerId;